Fixed a minor warning on 64bit platforms.
git-svn-id: http://yate.null.ro/svn/yate/trunk@626 acf43c95-373e-0410-b603-e72c3f656dc1

+++ engine/String.cpp
@@ -900,7 +900,7 @@ bool Regexp::compile()
     if (c_str() && !m_regexp) {
 	regex_t *data = (regex_t *) ::malloc(sizeof(regex_t));
 	if (!data) {
-	    Debug("Regexp",DebugFail,"malloc(%d) returned NULL!",sizeof(regex_t));
+	    Debug("Regexp",DebugFail,"malloc(%d) returned NULL!",(int)sizeof(regex_t));
 	    return false;
 	}
 	if (::regcomp(data,c_str(),m_flags)) {
